test(c7u): the fold, the writer, and the assertion row
tests/test_adr_records.c carries the acceptance: fold the same ADR twice and
the store's digest does not move; fold it on two machines that never spoke and
merging adds nothing. Plus the mapping field by field, a folded record proved
byte-identical to one the plain contract builds from the same facts (so a
reader needs no ADR vocabulary), a row whose instant cannot be read refused
rather than dated by guess, and a write with no record store failing closed
instead of falling back to replacing the row.

test_mcp asserts the client's view: two updates leave two decision records and
the superseded text is still readable. test_tool_surface carries the Phase 0
row from both directions — no memory argument arrives on the deprecated tool,
no ADR vocabulary leaves it.

ADR tests point HYP_CACHE_DIR at a throwaway directory; a suite that appends
to a real corpus is a suite with a side effect.

Every touched file passes -Wall -Wextra -Werror -fsyntax-only. Suite not yet
run.

+++ src/memory/adr_records.c
@@ -70,7 +70,7 @@ bool hyp_adr_timestamp_to_ms(const char *iso, int64_t *out_ms) {
     if (!iso || !out_ms) {
         return false;
     }
-    if (strlen(iso) != HYP_ADR_TIMESTAMP_LEN - 1) {
+    if (strlen(iso) != (size_t)(HYP_ADR_TIMESTAMP_LEN - 1)) {
         return false;
     }
     if (iso[4] != '-' || iso[7] != '-' || iso[10] != 'T' || iso[13] != ':' || iso[16] != ':' ||
@@ -110,7 +110,7 @@ bool hyp_adr_timestamp_to_ms(const char *iso, int64_t *out_ms) {
 }
 
 bool hyp_adr_timestamp_from_ms(int64_t ms, char *buf, size_t bufsz) {
-    if (!buf || bufsz < HYP_ADR_TIMESTAMP_LEN) {
+    if (!buf || bufsz < (size_t)HYP_ADR_TIMESTAMP_LEN) {
         return false;
     }
     if (ms < HYP_RECORD_MIN_TIMESTAMP_MS || ms > HYP_RECORD_MAX_TIMESTAMP_MS) {
